refactor: rename classes and packages to new optimization context
