+++ packages/route-optimization-jsprit/.gitignore
@@ -13,3 +13,4 @@ bin/
 
 ### Misc
 tmp/
+output/
