+++ packages/route-optimization-jsprit/src/test/java/com/predictivemovement/route/optimization/MQApplicationTests.java
@@ -1,10 +1,10 @@
-package com.predictivemovement.evaluation;
+package com.predictivemovement.route.optimization;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EvaluationApplicationTests {
+class MQApplicationTests {
 
 	@Test
 	void contextLoads() {
